Unify bracket/swiss score widths to i64
BracketMatch/SwissMatch scores were Option<i32> while TeamView/NormTeam,
the DB, and the PandaScore parse all use i64, so build_rounds/build_swiss
cast every score `as i32`. Widen the bracket/swiss score fields (and the
render structs / score_view that mirror them) to Option<i64> and drop the
casts — one score width end to end, no behaviour change (map counts are
tiny either way).

+++ src/types.rs
@@ -232,8 +232,8 @@ pub struct StandingRow {
 pub struct BracketMatch {
     pub team_a: String,
     pub team_b: String,
-    pub score_a: Option<i32>,
-    pub score_b: Option<i32>,
+    pub score_a: Option<i64>,
+    pub score_b: Option<i64>,
     /// Which side won, if decided: "a", "b", or empty.
     pub winner: String,
     /// PandaScore match id, for linking to the match detail page. 0 when unknown
@@ -254,8 +254,8 @@ pub struct BracketMatch {
 pub struct SwissMatch {
     pub team_a: String,
     pub team_b: String,
-    pub score_a: Option<i32>,
-    pub score_b: Option<i32>,
+    pub score_a: Option<i64>,
+    pub score_b: Option<i64>,
     /// Which side won, if decided: "a", "b", or empty.
     pub winner: String,
     pub match_id: i64,
